Tests/LibWeb/Text/expected/hit_testing/inline-element-with-inline-block-child.txt

<a> rect matches <i> rect: true
